+++ src/main/java/org/hv/pocket/criteria/PersistenceProxy.java
@@ -12,6 +12,7 @@ import org.hv.pocket.utils.EnumPocketThreadPool;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.Serializable;
 import java.lang.reflect.Field;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -78,7 +79,7 @@ public class PersistenceProxy {
             long milliseconds = System.currentTimeMillis() - startTime;
             this.consoleLog(sql, milliseconds);
             // 生成后镜像
-            if (this.databaseNodeConfig.getCollectLog() && this.showSqlLog) {
+            if (this.databaseNodeConfig.getCollectLog() && this.databaseNodeConfig.getShowSql()) {
                 this.pushLog(sql, null, null, milliseconds);
             }
         }
@@ -98,7 +99,7 @@ public class PersistenceProxy {
             long milliseconds = System.currentTimeMillis() - startTime;
             this.consoleLog(sql, milliseconds);
             // 生成后镜像
-            if (this.databaseNodeConfig.getCollectLog() && this.showSqlLog) {
+            if (this.databaseNodeConfig.getCollectLog() && this.databaseNodeConfig.getShowSql()) {
                 this.pushLog(sql, null, null, milliseconds);
             }
         }
@@ -147,7 +148,7 @@ public class PersistenceProxy {
             long milliseconds = System.currentTimeMillis() - startTime;
             this.consoleLog(sql, milliseconds);
             // 生成后镜像
-            if (this.databaseNodeConfig.getCollectLog() && this.showSqlLog && result > 0) {
+            if (this.databaseNodeConfig.getCollectLog() && this.databaseNodeConfig.getShowSql() && result > 0) {
                 this.pushLog(sql, beforeMirror, this.loadAfterMirror(beforeMirror), milliseconds);
             }
         }
@@ -182,7 +183,7 @@ public class PersistenceProxy {
     }
 
     private void consoleLog(String sql, long milliseconds) {
-        if (this.showSqlLog) {
+        if (this.databaseNodeConfig.getShowSql()) {
             this.executorService.execute(() -> logger.info("【SQL】 {} \n 【Milliseconds】: {}", sql, milliseconds));
         }
     }