+++ src/test/java/org/hv/dbconnect/QueryTest.java
@@ -114,4 +114,47 @@ public class QueryTest {
         List<OrderView> orders = query.list();
         System.out.println(orders.size());
     }
+
+    @Test
+    public void test8() throws SQLException {
+        SQLQuery query = this.session.createSQLQuery("delete from tbl_order where CODE = :ORDER_CODE AND DAY < :DAY", OrderView.class)
+                .setParameter("ORDER_CODE", "C-001")
+                .setParameter("DAY", new Date());
+        int row = query.execute();
+        System.out.println(row);
+    }
+
+    @Test
+    public void test9() throws SQLException {
+        SQLQuery queryInsert = this.session.createSQLQuery("insert into tbl_order(uuid,code,price) values('00001','C-00001', 10)");
+        int rowInsert = queryInsert.execute();
+        System.out.println(rowInsert);
+        SQLQuery queryUpdate = this.session.createSQLQuery("update tbl_order set price = 11 where uuid='00001'");
+        queryUpdate.execute();
+        int rowUpdate = queryUpdate.execute();
+        System.out.println(rowUpdate);
+        SQLQuery queryDelete = this.session.createSQLQuery("delete from tbl_order where uuid='00001'");
+        int rowDeleted = queryDelete.execute();
+        System.out.println(rowDeleted);
+    }
+
+    @Test
+    public void test10() throws SQLException {
+        SQLQuery queryInsert = this.session.createSQLQuery("insert into tbl_order(uuid,code,price) values(:UUID, :CODE, :PRICE)");
+        for (int index = 0; index < 10; index++) {
+            queryInsert.setParameter("UUID", "2020" + index)
+                    .setParameter("CODE", "C-00" + index)
+                    .setParameter("PRICE", index)
+                    .addBatch();
+        }
+        int[] rowInserts = queryInsert.executeBatch();
+        System.out.println(Arrays.toString(rowInserts));
+        SQLQuery queryDelete = this.session.createSQLQuery("delete from tbl_order where uuid = :UUID");
+        for (int index = 0; index < 10; index++) {
+            queryDelete.setParameter("UUID", "2020" + index)
+                    .addBatch();
+        }
+        int[] rowDeletes = queryDelete.executeBatch();
+        System.out.println(Arrays.toString(rowDeletes));
+    }
 }